# Avoid copy of intersection in totalTurnAngle